feat/90: Add Position column to facts table and update types
This commit introduces a new nullable integer column named `Position` to the `facts` table in the database schema. It also modifies the `StreetcodeType` column in the `streetcodes` table to have a maximum length of 21 characters, replacing the previous `nvarchar(max)` type. Similarly, the `CoordinateType` column in the `coordinates` table is updated to have a maximum length of 21 characters. These changes are part of a migration to ensure the database schema is optimized for data integrity and storage efficiency.