feat/214: Add RefreshToken entity and migration
Introduce a new `RefreshToken` entity to the database context of `UserService.WebApi`. This includes the creation of the `AddRefreshTokenEntity` migration class, which defines the `RefreshTokens` table structure, including columns and constraints. Set up foreign key relationships with the `AspNetUsers` table and create indexes for `Token` and `UserId`. Update the model snapshot to reflect these changes in the Entity Framework Core context.